9 Nov 2023: adding Utility to launch 64-bit Windows applications in 64-bit Wine on my system so that I don't have to remember the details of the invocation.
Usage:
    wine64.py EXECTUABLE_NAME [options for Windows program]

Uses unsafe behavior to be lazy about how programs are launched, and is
therefore not a good option to allow untrusted users to use.

This script is copyright 2023 by Patrick Mooney. It is licensed under the GNU
GPL, either version 3 or (at your option) any later version. See the file
LICENSE.md for details., a convenience script to run a Windows program on my system using the 64-bit installation of Wine.